+++ sky/shell/gpu/direct/ganesh_canvas.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/shell/gpu/ganesh_canvas.h"
+#include "sky/shell/gpu/direct/ganesh_canvas.h"
 
 #include "base/logging.h"
 #include "third_party/skia/include/gpu/gl/GrGLInterface.h"
